Added utility function to spawn service in specific group
Some services write fields that need to be readable/writable to
members of arbitrary groups users. It is unnecessary to create fake
users for all these cases (not all su implementation let one change just
group), so a small utility helper was written to do just that. setgid
spawns changes the gid and execs into a shell that runs the arbitrarily
command