Improved service management
There was a bug when service-manager would kill the service but not any
of its children. For the common case where the literal service is just a
shell file, than means any running program is left alive.

Also removed the respawn command and the dependence on /bin/pkill in
general.