service-spawn: Use O_CLOEXEC when opening pid file
The actual service used to inherit the file descriptor for the pid file.
And so did all of its decedents. This basically led to every program have
the pid file descriptor of the tty which is wasteful. Now the fd should
be closed on exec so only service-spawn should have it.